[Maha]: Fix Warning

+++ IR/IR/HashTableDictionary.cpp
@@ -180,7 +180,7 @@ unsigned long long HashTableDictionary::getTokenId(string token)
 Term * HashTableDictionary::getTermById(unsigned long long id)
 {
 	unsigned int index = id % size;
-	unsigned int i = id / size;
+	unsigned int i = (unsigned int) (id / size);
 	list<Term>* cell = hashTable[index];
 	if (cell != nullptr)
 	{
